+++ src/Yasumi/Provider/Germany/LowerSaxony.php
@@ -31,4 +31,38 @@ class LowerSaxony extends Germany
      * country or sub-region.
      */
     const ID = 'DE-NI';
+
+    /**
+     * Initialize holidays for Lower Saxony (Germany).
+     *
+     * @throws \Yasumi\Exception\InvalidDateException
+     * @throws \InvalidArgumentException
+     * @throws \Yasumi\Exception\UnknownLocaleException
+     * @throws \Exception
+     */
+    public function initialize()
+    {
+        parent::initialize();
+
+        // Add custom Christian holidays
+        $this->calculateReformationDay();
+    }
+
+    /**
+     * For the German state of Lower Saxony, Reformation Day is celebrated since 2018.
+     * Note: In 2017 all German states will celebrate Reformation Day for its 500th anniversary.
+     *
+     * @link https://www.zeit.de/gesellschaft/zeitgeschehen/2018-06/reformationstag-niedersachsen-neuer-feiertag
+     *
+     * @throws \Yasumi\Exception\InvalidDateException
+     * @throws \InvalidArgumentException
+     * @throws \Yasumi\Exception\UnknownLocaleException
+     */
+    private function calculateReformationDay()
+    {
+        if ($this->year < 2018) {
+            return;
+        }
+        $this->addHoliday($this->reformationDay($this->year, $this->timezone, $this->locale));
+    }
 }